requirements.txt
django==3.1.2
djangorestframework==3.11.1
django-filter==2.4.0
drf-yasg==1.17.1

django-redis==4.12.1
redis==3.5.3

pyyaml==5.3.1
psycopg2==2.8.6
